+++ Mods ChoGGi/Dome Show Trait Filters/metadata.lua
@@ -1,18 +1,19 @@
 return PlaceObj("ModDef", {
 	"title", "Dome Show Trait Filters",
-	"version", 1,
-	"version_major", 0,
-	"version_minor", 1,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_DomeShowTraitFilters",
-	"steam_id", "1817670791",
 	"pops_any_uuid", "712190e9-7b61-415d-9c64-7a33195426f5",
-	"author", "ChoGGi",
+	"steam_id", "1817670791",
 	"lua_revision", 1007000, -- Picard
+	"version", 2,
+	"version_major", 0,
+	"version_minor", 2,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
 	"TagInterface", true,
-	"description", [[Show the trait filters set for a dome in the traits button tooltip.]],
+	"description", [[
+Show the trait filters set for a dome in the traits button tooltip.
+]],
 })