+++ Mods ChoGGi/Stop Rocket Camera/Code/Script.lua
@@ -21,7 +21,7 @@ function OnScreenNotification:CycleObjs(cycle_objs, ...)
 		return ChoOrig_CycleObjs(self, cycle_objs, ...)
 	end
 
-	if cycle_objs and self.notification_id == ("NewTradeOffer" .. GetGameMap(self).map_id) then
+	if cycle_objs and self.notification_id == ("NewTradeOffer" .. UICity.map_id) then
 		return cycle_objs[1]
 	end
 	return ChoOrig_CycleObjs(self, cycle_objs, ...)
@@ -32,14 +32,14 @@ local ChoOrig_ViewAndSelectObject = ViewAndSelectObject
 local fake_ViewAndSelectObject = empty_func
 
 local ChoOrig_SendRocketToMarsPoint = SendRocketToMarsPoint
-function SendRocketToMarsPoint(...)
+function SendRocketToMarsPoint(obj, spot, ...)
 	if spot.spot_type == "project" then
-		return ChoOrig_SendRocketToMarsPoint(...)
+		return ChoOrig_SendRocketToMarsPoint(obj, spot, ...)
 	end
 
 	ViewAndSelectObject = fake_ViewAndSelectObject
 
-	ChoOrig_SendRocketToMarsPoint(...)
+	ChoOrig_SendRocketToMarsPoint(obj, spot, ...)
 	-- we need the WaitMsg since the orig func does it as well
 	CreateRealTimeThread(function()
 		WaitMsg("PlanetCameraSet")