+++ Mods ChoGGi/Stop Rocket Camera/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "1796377374",
 	"pops_any_uuid", "7af8a51c-f8ea-4528-aee9-e42368fee80d",
 	"lua_revision", 1007000, -- Picard
-	"version", 4,
+	"version", 5,
 	"version_major", 0,
-	"version_minor", 4,
+	"version_minor", 5,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {